se modificó el modelo, se hicieron correcciones con codigo en python 3.7 y se mejora la legibilidad y tambien reutilizacion de la conexion a la base de datos

+++ formularios.py
@@ -1,10 +1,10 @@
 from flask_wtf import FlaskForm
-from wtforms import IntegerField, SelectField, StringField, DecimalField, SubmitField
+from wtforms import FloatField, IntegerField, SelectField, StringField, DecimalField, SubmitField
 from wtforms.validators import DataRequired, NumberRange
 
 class ProductoForm(FlaskForm):
     marca = StringField('Marca', validators=[DataRequired()])
-    precio = DecimalField('Precio', validators=[DataRequired(), NumberRange(min=0)])
+    precio = FloatField('Precio', validators=[DataRequired(), NumberRange(min=0)])
     submit = SubmitField('Crear Producto')
 
 class VentaForm(FlaskForm):
